requirements.txt
anthropic==0.49.0
twilio==9.4.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
